[NONE] Fix issues in Line#contains(Point) and Polygon#contains(Point).
The Line#contains(Point) method has to test its end points for equality
with the given point in question in addition to the original test.
Otherwise, the distance to the straight through the line can possibly
dismiss the point in question, although it is equal to one of the end
points.

The Polygon#contains(Point) method has to test if the intersection point
calculated during the ray shooting is on the side of the ray, relative
to the point in question. Otherwise, a point very near to an edge of the
polygon may be dismissed, because the intersection is falsely counted.