[NONE] Enhance an exception message in Ring.
The triangulation methods in the Ring class can throw an
IllegalStateException if the algorithm fails at determining points of
intersection of a triangle and a line. This exception did not contain
information on the parameter values. To be able to reproduce the error,
such information is added to the exception message.