[NONE] Ensure Arc#equals(Arc) uses Angle#equals(Angle).
The Arc#equals(Arc) method used PrecisionUtils to compare the start
angle and the angular extent of both objects. This behavior is incorrect
for angles around the 0deg/360deg border that are imprecisely equal.

The Angle#equals(Angle) method implements the correct, imprecise
comparison for Angle objects. This fix makes use of the
Angle#equals(Angle) method in the Arc#equals(Arc) method.

Additionally, a test case for the false behavior is added to the test
suite.