[384328] Ensure that Vector objects instead of Point objects are used in Straight#getPointAt() and Straight#getParameterAt()
The getPointAt() and getParametetAt() methods in the Straight class are
inconsistent to the rest of the API in that they are working with Point
objects instead of Vector object.

This commit fixes this inconsistency by adjusting the methods to work
with Vector objects instead of Point objects. Additionally, the
getPointAt() method is renamed to getPositionVectorAt() so that its
return type is represented by its name. All invocations of the prior
methods are adjusted to support the new interface.

Furthermore, the error handling of the getParameterAt() method in the
Straight class is changed. If the input contract for that method gets
discounted, i.e. the passed-in Vector does not lie on the Straight, an
IllegalArgumentException is thrown where previously 0 was returned which
is a valid return value in the non-exceptional case.