[NONE] Align the BezierCurve#getParameterAt() exceptional behavior with that of Straight#getParameterAt().
The exceptional behavior of the BezierCurve#getParameterAt() method is
modified to throw an IllegalArgumentException if the passed-in Point
does not lie on the BezierCurve.

This is the same behavior is in the Straight#getParameterAt() method.

Fix BezierCurve#getParameterAt() test case.

In the old test case for that method, a NullPointerException is expected.
But the changes to getParameterAt(Point) require to catch an
IllegalArgumentException insteaumentException instead.