[383647] Implement Ring#toSWTRegion() using getOutline().
The toSWTRegion() method did not account for the SWT +1/-1 pixel
behavior when using a simple (int) cast of the double coordinates of the
interval Polygons.

The new getOutline() method returns closed Polylines for each closed
outline. Therefore, using this method to build the SWT Region does not
have the +1/-1 pixel problems.