[NONE] Create CurvedPolygon.
Adds another IShape implementation to the GEF4 Geometry API. The
CurvedPolygon encloses an area using BezierCurve edges. The outline has
to be consecutive within the imprecision used throughout the whole API.
Internally, the outline segments are modified so that the common end
points of two consecutive BezierCurves are identical.

A simple test for the CurvedPolygon#contains(Point) method is added to
the test suit and a utility method to copy a bunch of BezierCurves is
added to the CurveUtils class.

Does also add CurvedPolygonTests.class to AllTests.java.