[NONE] Refactor parts of examples.
A new set of utility classes is added to the examples package to reduce
the amount of code required to write a simple example application that
demonstrates functionality of the GEF4 Geometry API.

Some of the already existing example applications are updated to use
these new utility classes, i.e. subclass the AbstractExample class.
Additionally, a few new example applications are provided which build
upon these new utility classes.