[384059] Ensure Point#getCentroid() computes correct coordinate values for the imaginary closing segment.
The computation of the centroid is based on assuming that the points are
forming a polygon of constant density. The x and y coordinates are
iteratively computed for each segment of that polygon. A typo in the
computation of the coordiante values for the closing segment led to
false values there.