[NONE] Implement CAG functionality in Path.
Extend the Path class by 4 static methods that do CAG operations:
add(), intersect(), exclusiveOr(), and subtract().

Add an example application to demonstrate the intersect() operation on
a Polygon and an Ellipse. The ControllableShapeViewer class is modified
so that the control points of the ControllableShapes are drawn after
the ControllableShape#onDraw() method is called on each of them. This
ensures, that the control points are always in front of the other items
on the screen.