[NONE] Implement AbstractPolyShape#toPath().
The Region#toPath() and the Ring#toPath() methods have been removed. Instead, a
general AbstractPolyShape#toPath() method is added. Furthermore, the new
toPath() method uses the getOutline() method and CAG functionality to construct
a correct Path, which was not the case before.

Tests are added to verify that the toPath() method is working correctly, even
when dealing with voids in an IPolyShape.