[388673] Fix AffineTransform#equals(Object).
Prior to this fix, the AffineTransform#equals(Object) method called the
java.awt.geom.AffineTransform#equals(Object) method to check for equality. But
it passed an object of wrong type to that method, so that it did always return
false.

This fix corrects the delegation by passing in a java.awt.geom.AffineTransform
representation of the given GEF4 AffineTransform.