[386029] Add missing IGraphicsProperties tests.
All five IGraphicsProperties are now tested for SWT ang AWT DisplayGraphics
objects. The tests are restructured to not use the 'Enclosed' JUnit Runner which
causes the Eclipse JUnit View to display individual tests incorrectly (not
checkmarked) in this situation, and which is unnecessary anyway.

The accessor methods in the IGraphics interface for the individual
IGraphicsProperties are renamed to not include the 'get' at the start of the
method names. This is done to preserve the convention that a method whose name
starts with 'get', returns a copy of the actual object. So as these methods do
return the actual object associated with the IGraphics, the accessor method
names should not start with 'get'.