[386029] Add GradientStop normalization in SWTFillProperties.
In order to be able to delegate the gradient drawing to SWT, a start and an end
GradientStop are inserted into the list of GradientStops if the first
user-specified GradientStop is not at a percentual distance of 0% (0.00), or the
last user-specified GradientStop is not at a percentual distance of 100% (1.00),
respectively.

SWT does only support linear single stop gradients, i.e. there may
only be a start and an end GradientStop. Therefore, if the number of
GradientStops equals 2 after normalizing, we can create an SWT Pattern
representation of the gradient.

Eventually, the normalization code will move to the GradientFill class in a
future version.