[390172] Disable dynamic layout by default
This restores the Zest 1.x behavior where changing the graph
structure and refreshing the viewer only shows the structural
change, but does not apply the layout algorithm.

Dynamic layout is enabled automatically when an
ExpandCollapseManager is added to the graph and can be enabled
manually for custom usage (see SpringLayoutProgress example).