[386029] Refactor GEF4 Graphics abstractions and the AWT implementations.
The GEF4 Graphics-component is restructured to provide direct access to the
properties which control the individual drawing operations. The prior
IGraphicsProperties, its descendents, and implementations were over-engineered
and complicated the use of the component.

Moreover, packages have been renamed and classes have been moved to reflect
their relations. However, some names are not yet pleasant and may undergo
changes in the future.