[386029] Add emulate-xor-mode flag to IGraphics.
Xor-mode rendering is emulated when using SWT on GTK as gc.setXORMode(true)
does not work correctly on GTK if any of the "advanced" GC features is used. The
user can choose to always emulate xor-mode rendering even if not on GTK by
setting the new emulate-xor-mode flag to true.