[410772] Fix location translation in coordinate transformations.
The various methods in the AbstractNode, AbstractParent, ControlNode, and Pane
classes to transform coordinates from one system into another have to take the
absolute location of the Scene (SWT Control) into consideration.

In addition, the computation of layout-bounds of an AbstractParent is modified
to contain all transformed bounds (bounds-in-parent) of its children. Therefore,
the Pane class does now override getLayoutBounds() and the Group class does not
anymore.