[410772] Fix transparency issue.
The global-alpha attribute of the GraphicsContext class was incorrectly handled
through-out the code. Its value is an integer in range [0;255] where 0 signifies
full transparency and 255 signifies full opacity. Some code expected the value
to be a double in range [0;1].

In the process of fixing the issue, a helper function WITH_ALPHA() was added to
GraphicsContext. WITH_ALPHA receives a local alpha value and a Runnable. It
combines local and global alpha on the underlying GC, calls the Runnable, and
resets the GC's alpha afterwards.