[410772] Simplify querying mouse event coordintes.
A MouseEvent object does now provide methods to query the mouse coordinates
relative to the event's target, the Scene in which the event occured, or the
current Display. One field is used for each of the attributes. The
SwtEventForwarder is responsible for filling-in the correct values; coordinate
transformations are performed there.