[410772] Fix ConcurrentModificationException in PulseThread.
The PulseThread manages a list of IPulseListeners which are informed on every
application pulse. Animations may well be modifying that list during execution
to deregister themselves upon completion. Therefore, the loop informing the
listeners is vulnerable to ConcurrentModificationExceptions. I fixed the issue
by looping over a copy of the current state of listeners.