[NONE] Move anchorage reference point computation from anchor to anchored.
Previously an IAnchor was responsible for computing the reference point used to
determine its real anchor point. This responsibility is now carried by the
anchoreds.

Moreover, this commit introduces a change in visual registration and
synchronization order which was necessary in order to be able to register layout
listeners on the anchorages.