[NONE] Introduce IBehavior (and migrate non-tooling related IPolicys).
- Introduced IBehavior as a complement to IPolicy and IPartBound as
super-interface, which is used for registering policies and behaviors at
parts. While policies are meant as an interface for tools (and are thus
no longer activatable by default, behaviors is independent of tools (but
activatable). 
- Migrated feedback and synchronization policies into behaviors, as they
are not accessed by tools, but are registered locally and listen to
respective viewer models.
- Introduced IViewerBound and IDomainBound interfaces in analogy to
IPartBound interface to formalize relationship.
- Replaced RelocateSelectedTool/FXRelocateSelectedTool with
AbstractRelocateSelectedOnDragPolicy/FXRelocateSelectedOnDragPolicy.