[NONE] Remove interpolation failure exception.
When interpolating cubic bezier curves through a given set of anchor points
(way points) the resulting beziers were checked for completeness. In case not
enough beziers were created an IllegalStateException was thrown.

The check and the exception are removed because the interpolation algorithm
assures that the returned array of BezierCurves is of correct length.

The check and the exception have been inserted in the first place, because we
observed a bug which seemed to be related to interpolation but proved to have
its origin in other reasons.