[430940] Add filter to suppress mouse wheel events on touch devices.
Added an SWT event filter to suppress mouse wheel events on devices that
support touch (Display#getTouchEnabled()). While on Mac (where I could
observe the synthetic events and could test the correct behavior with
the applied filter), only vertical mouse wheel events seem to be
transferred to the FXCanvas, I added the suppress filter also for
horizontal mouse wheel events just in case this occurs on other
platforms as well (which I cannot confirm). We will have to test that
the behavior is valid on other platforms than Mac to see if the behavior
has to be adjusted in a platform-specific way.