[429511] Refactor doc generation process.
Centralized generation of documentation into org.eclipse.gef4.releng
project. It now contains the libs, as well as two ant scripts
(generateWikitext.xml and generateJavadoc.xml) that are called from the
org.eclipe.gef4.releng/pom.xml and generate documentation in case a
javadocOptions.txt and a wikitextOptions.txt file can be found in the
project.basesdir. Changed the doc plug-ins to no longer bundle the libs
and the generateDoc.xml ant script, but instead (in addition to the
already existing javadocOptions.txt) to provide only a
wikitextOptions.txt file.