[NONE] Implement AbstractFXColorPicker control.
The control consists of two elements: a rectangular area to display the
currently selected color, and a button with a color wheel. The control is
abstract, because you have to complement it with a dialog to choose a new color.
This dialog is opened when the user clicks the button.

The JavaFX ColorPicker contains such a dialog. If we can open it independent of
the ColorPicker itself, this would be a good default implementation.