[NONE] Major refactoring of GEF4 MVC introducing DI.
- Introduce dependency injection based on Google Guice. This includes
polymorhphic injection of adapters into IAdaptables (by using a
dedicated AdapterMap binding annotation, and a type listener and member
injector to evaluate the polymorphic bindings).
- Changed that ITools are now registered as adapters on the IDomain;
merging FXClickTool and FXDragTool to a single FXClickDragTool, as click
policies will have to be notified before drag policies (and ordering of
adapters may not be relied on).
- Ensured an IDomain can hold references to multiple IViewers, not just
a single one. The viewer models (SelectionModel, HoverModel, etc.) are
bound to a specific viewer, whereas the ITools (bound to the IDomain)
have to be able to work with several viewers. Removed IDomainBound and
IViewerBound interfaces.
- Renamed IFXSceneHook to ISceneContainer and removed it from
Constructor of FXViewer, so it can be passed into the viewer (and the
class can be properly constructed via an injector) lazily.
- Updated build scripts to use Luna target by default; added Guice
bundles to Luna target.
- Ensured PrecisionUtils are safe-guarded against comparison of
Double.NaN values.