[438369] Extend anchor mechanism to support multiple positions for the same anchored.
An anchor position is no longer assigned per anchored Node only, but an
arbitrary identifier is additionally used to differentiate multiple positions
for the same anchored.

A combination of anchored Node and arbitrary (Object) identifier is called an
AnchorKey. The combination of an IFXAnchor and an AnchorKey is called an
AnchorLink. Therefore, an AnchorLink represents exactly one anchor position.