[NONE] Fix disappearing visuals when zoomed and interacting.
In (seemingly) arbitrary situations, the visuals of the GEF4 MVC example
disappeared when the content was scaled (zoomed) and mouse events were
processed, for example, when dragging a node.

This is now fixed by the (seemingly) arbitrary registration of a ChangeListener
for the boundsInLocalProperty of the content layer. The registration of the
listener leads to a different ExpressionHelper being used by JavaFX to manage
the boundsInLocalProperty. That ExpressionHelper calls
boundsInLocalProperty#getValue() when a change event is fired, which revalidates
(recomputes) the bounds.