[NONE] Downgrade e(fx)clipse to Kepler release until 1.0.0 is finalized.
Up to now, e(fx)clipse was a moving target, as the 1.0.0 version can
only be obtained from the nightly e(fx)clipse update-site. Changing this
to 0.9.0 site allows to keep the version stable. e(fx)clipse 1.0.0
release is scheduled for July 31st, so we may switch to the release
version then.