[NONE] Fix reference point computation when no way points can be used.
When no way points are specified or suitable for computing the reference points
for FXChopBoxAnchor, the center point of the start/end anchorage is used as the
reference point.

The center was computed in scene coordinates. This has been changed to local
anchored coordinates as scene coordinates change when advancing the scrollable
area (top or left).