[NONE] Ignore mouse exit target events for determining mouse release.
It may happen that as a result of the preceding press event, a mouse
exit target event is triggered (if for instance the visual refresh that
is triggered by the press removes the childen like done in
FXCurvePart#refreshVisuals()). We need to ensure that these events are
ignored when determining the release.