[450223] Remove FocusPolicy, HoverPolicy, ZoomPolicy, and SelectionPolicy.
Removed those policies, which only manipulated the interaction models. Inlined the code within those interaction policies that delegated to them before. With the exception of the FXFocusAndSelectOnClickPolicy, the code is a simple delegation to an interaction model (there it could be refactored to be, by moving the selection logic completely to the selection model).