[NONE] Fix content removal.
When removing a content part, its anchor relations to other parts are detached
before the part is removed. When removing multiple parts, all of them have to be
detached before removing their contents. This is now done by delayed the
detachment up to the point where the object really is removed.