[NONE] Centralize pool for deleted content parts in a GraveyardModel.
Prior, every ContentBehavior managed its own content part pool. But if a
content part was removed using a different ContentBehavior than for a later
re-creation, the part could not be found in the local pool of the other
ContentBehavior, and therefore, a new part was created, invalidating all
existing references to the old part.