[NONE] Restructure policies so that resizing and transforming are the basic operations.
The FXResizePolicy is responsible for changing the size of a part. For example,
if you have a 45deg rotated rectangle, and you change its width, it will not
only grow horizontally, but also vertically, i.e. it depends on the orientation.

The FXTransformPolicy is responsible for transforming a part, i.e. scale,
rotate, shear, or translate it. Previously a relocation adjusted the layout-x
and layout-y properties of the corresponding node. Now, the relocation is
carried out by a translation transformation.

Other policies (ResizeRelocate, ScaleRelocate, RROnHandleDrag, SROnHandleDrag,
etc.) are based on these two policies. Therefore, it is sufficient for an
application to extend these two policies in order to write changes to the
underlying model.