[NONE] Fix replacing contents with another nested content element.
When replacing contents, the removal of no longer applicable parts has to be
done before the creation of new parts, because previously nested content
elements could be used as "new" content elements, leading to an
IllegalStateException.

For example, consider the following situation:

	Tree a = new Tree()
	Tree b = new Tree().setLeft(a)
	setContents(b)
	setContents(a)

If removal was done after creation, the second setContents() call would throw
an exception, because it would find that there already is a content part for
Tree a.