[461421] Do not use Layout properties to maintain Zest state.
GEF4 Zest is able to hide nodes. Hidden nodes are not taken into account during
layout. The "hidden" state was previously maintained as a Layout property,
although it is unspecific to GEF4 Layout. That's why the "hidden" state is now
maintained as an attribute on the Zest graph. A filter mechanism has been re-
introduced to be able to disregard hidden nodes from layout.