[461297] Automatically refresh upon Graph changes (attributes, nodes, edges).
The individual content parts (NodeContentPart, EdgeContentPart,
GraphContentPart) register property change listeners on their content objects
(Node, Edge, and Graph, respectively).

When the attributes of a Node, Edge, or Graph change, the corresponding visual
is refreshed. When the nodes or edges of a Graph change, the content parts are
synchronized with the content objects.

Besides, many HashSets have been replaced with Collections.newSetFromMap(new
IdentityHashMap()) so that nodes/edges are identified by their object identity.