[477352] Refactor FXTransformPolicy and all users.
The FXTransformPolicy does now manage a list of pre-transforms and a list of
post-transforms. The user can create pre- and post-transforms and manipulate
them later. The final transformation is computed by concatenating all pre-
transforms, then the original node transform, and then all post-transforms.

The intermediate FXResizeRelocatePolicy, FXScaleRelocatePolicy, and
FXRotatePolicy have been removed. All policies that were based on these
intermediate policies have been updated to use the FXTransformPolicy directly.

In some parts, the implementations of the other related policies have been
outsourced to the FXTransformPolicy, such as a translation in scene
coordinates, a rotation around a pivot point (in scene coordinates), and the
scaling around a pivot point (in scene coordinates).