[471154] Refactor color and gradient pickers and related cell editor.
- Merged AbstractFXColorPicker and nested ColorWheel into FXColorPicker.
Changed FXColorPicker to be an SWT Composite and refactored it
internally.
- Renamed FXSimpleGradientPicker into FXSimpleLinearGradientPicker,
changed it to be an SWT Composite, and refactored it internally.
- Renamed FXAdvancedGradientPicker into FXAdvancedLinearGradientPicker,
changed it to be an SWT Composite, and refactored it internally.
- Increased versions of MVC.FX.UI bundle and feature to 0.3.0. In
consequence adjusted versions of MVC example and doc bundles, as well as
enclosing features to 0.3.0.
- Increased versions of clients bundles and features (Zest.FX.UI,
DOT.UI) and related test, examples, and sdk bundles/features to 0.3.0.