[470636] Extract Zest.FX.UI JFace-API into own Zest module.
- Created Zest.FX.JFace module (org.eclipse.gef4.zest.fx.jface bundle
and related org.eclipse.gef4.zest.fx.jface-feature).
- Renamed o.e.g4.zest.fx.examples.ui into o.e.g4.zest.fx.examples.jface,
as all provided examples are related to JFace-API.
- Moved ICanvasFactory from MVC.FX.UI to FX.SWT (it is not related to
MVC) and ensured that Zest.JFace module uses it as well to create the
canvas. Ensure that an SWT style can be passed ot the created canvas.
- Defined an own ZestFxJFace module as an extension to ZestFxModule,
which properly binds an ICanvasFactory to create an FXCanvasEx.
- Reimplemented selection handling inside ZestContentViewer (so its no
longer based on SelectionForwarder and follows the JFace contract).