[478408] Introduce ReadOnlyMapWrapperEx to fix a JavaFX bug.
The JavaFX MapProperty allows the registration of change listeners. However,
there is a JavaFX bug that causes all change listeners to be removed when a
single change listener is removed.

As a workaround, a custom implementation of ReadOnlyMapWrapper keeps track of
the listeners and ensures that the remaining listeners are still registered
after one listener was removed.