[482972] Make adapter registration and retrieval more intuitive.
-Changed IAdaptable interface to provide actual type information about
the to be registered adapter in case the adapter type is now raw type
(and it can thus not be properly determined because of type erasure).
-Changed that unregistering of adapters is performed through the adapter
not through the key.
-Changed implementation of AdaptableSupport to enable that adapters can
be retrieved by any type key in between the registration type key and
the actual type of the adapter.
- Adjusted AdapterStore to changes in IAdaptable and AdaptableSupport.
Removed setAdapters() method with Map argument and marked
setAdapter(AdapterKey, TypeToken, Object) as to be injected instead.
-Adjusted adapter map injection support to support injection of
setAdapter(AdapterKey, TypeToken, Object) method (and to infer the
approriate actual type of the adapter from the module bindings (even if
its a parameterized type).
-Adjusted AbstractDomain, AbstractViewer, and AbstractVisualPart to the
changes in IAdaptable and AdaptableSupport. Removed setAdapters() method
with Map argument and marked setAdapter(AdapterKey, TypeToken, Object)
as to be injected instead.