[482972] Ensure adapters are registered via their actual types.
- Changed setters in IAdaptable to no longer accept an AdapterKey (which
is only used for retrieving adapters), but to always register an adapter
with actual type only (in case its a parameterized type, that cannot be
inferred from the adapter instance).
- Changed getters in IAdaptable to comply to the platform's
IAdaptable#getAdapter(), i.e. instead of (getAdapter(Class<? super T>))
we now have getAdapter(Class<T>)  as well.
- Adjusted AdaptableSupport, AdapterStore, AbstractVisualPart,
AbstractDomain, and AbstractViewer to comply to the new contract.
- Adjusted all direct clients that registered and retrieved adapters to
comply to the changes in IAdaptable.
- Introduced a Types utility class (which supports to infer the actual
parameter type from a context class) and changed all adapter retrievals
within MVC to be type-safe, even if the VR parameter is unbound.
- Changed that adapter map bindings may now omit the type key within the
used AdapterKey in case it can be inferred from the adapter instance
itself, or through the binding (i.e. in case  a constructor binding is
used). AdapterKey contract had to be changed to allow this (it now
tolerates null keys).
- Enhanced AdapterMapInjector to handle AdapterKeys without type keys
(and to infer the actual adapter type from the bindings, or the
instance).
- Enhanced AdaptableTypeListener to perform some sanity checks on
potential injection points. That is, in order to not collide with Guice
member injection, the adapter map injection method may not also specify
an inject annotation. Further, the adapter map annotation may not
specify an adaptable type.