[483568] Remove IConnectionDecoration, properly align decorations.
- Changed that the local coordinate system of a decoration visual is
regarded as being 'anchored' at the start/end point of the connection.
Added a compensation to connection to ensure that decoration stroke is
compensated, by compensating the layout bounds min x. Due to this, the
start and end location of decorations is no longer meaniningful (we
expect it to be (0,0)-based). The
IConnectionDecoration abstraction is thus no longer of relevance either
and I removed it accordingly. Added a change listener to connection to
react to layout bounds changes of the decorations.
- Adjusted all IConnectionDecoration implementations to be simple Node
visuals. Ensured that Polygon instead of Polyline is used in order to
prevent line cap artifacts.