[443954] Fix GeometryNode layout bounds did not include stroke.
Ensured that the layout bounds of GeometryNode do include the stroke (as
in all other shapes). Ensured that when resizing a node, the geometric
bounds are re-calculated so that the resulting layout bounds (including
stroke) are matched (at least as far as possible). As a slight offset
may still remain, ensured that the returned layout bounds match the
values passed in to resize(). For this to work, in addition to
overwriting impl_ComputeLayoutBounds(), we now also have to call
impl_layoutBoundsChanged(), which is also deprecated.