[484774] Add property support for Multiset and SetMultimap.
- Add ObservableSetMultimap and related SetMultimapChangeListener, as
well as SetMultimapChangeListenerHelper support class. Add
ObservableMultiset and related MultimapChangeListener, as well as
MultisetChangeListenerHelper support class.
- Add ObservableSetMultimapWrapper and ObservableMultisetWrapper, which
can be used to easily create
ObservableSetMultimap and ObservableMultiset instances.
- Add SimpleSetMultimapProperty and ReadOnlySetMultimapWrapper, as well
as SimpleMultisetProperty and ReadOnlyMultisetWrapper to
realize writable and read-only property support.
- Moved ReadOnlyMapPropertyEx to org.eclipse.gef4.common.beans.property,
changed it to also fix problems related to unbinding of
BidirectionalBindings related to it, and added ReadOnlySetPropertyEx and
ReadOnlyListPropertyEx to fix the same problem.
- Added SimpleMapPropertyEx to fix the listener issue also holding for
ReadOnlyMapPropertyEx.
- Add test cases for ObservableSetMultimap and ObservableMultiset,
related properties, and for replacement classes.
- Add UnmodifiableMultimapWrapper and UnmodifiableSetMultimapWrapper.
- Add Bindings2 and extract logic from properties.