[486626] Only register "main" visuals at the visual-part-map.
The visual hierarchy can change on-the-fly, therefore, having all visuals
registered requires observing the visual hierarchy. It is not decidable which
visuals belong to which IVisualParts, because the visuals are hierarchically
organized, too. When registering the "main" visual of each IVisualPart
(#getVisual()), then the IVisualPart that controls a given visual can easily be
determined by walking up the visual hierarchy until a visual is found that is
registered for an IVisualPart.

- Removed registration of nested visuals.
- Centralized retrieval of IVisualPart for a given visual within FXPartUtils.