[489793] Ensure name and id are properly handled by DotInterpreter.
- Ensured that graph, node, and edge names are properly inferred by the
interpreter. 
- Ensured that graph, node, and edge id are properly handled as well.
- Refactored DotAttributes to only use a single constant to define the
respective attribute key. Used a respective '__GNE', '__G', '__NE' etc.
suffix to indicate to which elements the attribute applies. Ensured that
keys of attributes that do not refer to dot attributes, but to build-in
properties (name, graph type, etc.) are marked by a '_' prefix, as in
'_NAME__GNE', which is inferred from the ID value of the respective
graph, or node_id grammar rule (or computed for edges).