[488358] Fix segment overlay detection error in FXBendConnectionPolicy.
The segment overlay detection makes anchors explicit before determining if a
segment overlay is present. When it is determined that no overlay is present,
the anchors that were made explicit have to be made implicit again by removing
them from the explicit anchors. The errorneous code replaced them in the list
of explicit anchors with the previous implicit anchors.