[447980] Add arrowtype validation rules and perform minor adjustments.
- Add validation rule for checking the open modifier "o" in combination
with
invalid primitive arrow shapes.
- Modify validation rule for checking the side modifier "l"/"r" in
combination
with invalid primitive arrow shapes(the INV primitive arrow shape is
valid to combine with the side modifier - the Graphviz documentation is
somehow inconsistent).
- Add validation rule for checking that "none" is not the last arrow
shape, since this would create a redundant shape.
- Modify the DotJavaValidator to invoke the DotArrowTypeJavaValidator
also on the root element (not only on its children), otherwise the
DotJavaValidator check method with the ArrowType parameter is not
invoked by the Xtext framework.
- Extend the arrowshapes_multiple.dot sample file with "arrowtail",
"arrowsize" and "dir" attributes.
- Add arrowshapes_direction_both.dot sample file.
- Add arrowshapes_invalid_modifieres sample file.
- Implement additional DotParser and DotValidator test cases.
- Dot2ZestGraphConverter: use ".equals(...)" instead of "==" to compare
strings.
- DotArrowType.xtext: remove unnecessary parenthesis.

Signed-off-by: Tamas Miklossy <miklossy@itemis.de>
Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=477980