[441352] Preparations to support custom node shapes in Zest.
- Refactored DynamicAnchor-IComputationStrategy-contract, so that the
anchor is now responsible of providing the anchorage reference geometry.
Thereby, a specific anchor can be returned to ensure that anchors are
properly placed on the outline of the shape, whereas the computation
strategy can independently thereof by adjusted to match the connection
router.
- Added a specific anchor provider for NodePart, which refers to the
outline of the shape.
- Renamed some ZestProperties to correspond to the common conventions.
- Introduced a DotNodePart and corresponding DotContentPartFactory to be
able to exchange the dot shape. DotNodePart uses an ellipse as shape, as
this corresponds to the Graphviz default.
- Added a specific DotStraightRouter, which similar to the other Dot
routers ensures the the start and end point computed by Graphviz dot are
set as reference points for the start and end anchors.