[NONE] Fix BezierCurve#getIntersections() error.
The BezierCurve#getIntersections() method refines interval pairs to find
precise points of intersections between two bezier curves. If an interval is
split, then an epsilon is added to the second interval's lower boundary.
However, if the lower boundary is already close to 1, it might exceed the
maximum value of 1. Therefore, I added a check that prevents the addition of
the epsilon in such cases.