[495112] Remove DynamicAnchorELetterSnippet.
- Removed the DynamicAnchorELetterSnippet. As the logic of the
ChopBoxStrategy has been very much simplified (it now falls back to the
nearest projection in case the center is not contained in the anchorage
reference geometry), no explicit example is required to explain its
computation. We could think of building up a similar example but where
the center is contained.
- Removed it from the FX.Examples documentation as well as from the FX
documentation.