[498324] Improve the Dot Xtext editor content assistant support.
- Extend the GenerateDot.mwe2 workflow to generate content assistant
support for the dot sub-grammars: DotArrowTypeProposalProvider,
DotPointProposalProvider, DotShapeProposalProvider,
DotSplineTypeProposalProvider, DotStyleProposalProvider.

- Extend the DotProposalProvider class with the content assistant logic
for proposing the valid attribute names and the valid attribute values
considering the current context. Add the DotProposalProviderDelegator
class to delegate the content assistant logic to the proposal provider
of the corresponding sub-grammar.

- Implement corresponding test cases within the DotContentAssistTest
class (should be executed as JUnit Plug-in Test).