[501056] Remove compile-time dependency to JDK internals from FXCanvasEx.
Ensured TKSceneListener is accessed via reflection code only, and only
in a JavaSE-1.8 environment, where the workaround for JDK-8143596 is
still required. This is required because in a JavaSE-1.9 environment,
JIGSAW would prevent access to TKSceneListener (as the workaround is no
longer needed there, this should now be Java-9 compatible).