[517356] Replace IViewerDependent with IAdaptable.Bound<IViewer>.
- Ensured that AbstractViewerAction, AbstractViewerContributionItem, and
AbstractViewerActionGroup are IAdaptable.Bound<IViewer> at the content
viewer.
- Removed IViewerAction abstraction, which is not needed.
- Ensured that AbstractFXEditor creates and disposes the global delete
and selectAll actions, and that the FXEditorActionBarContributor is only
responsible for setting the global action handler to the specific action
implementation of the editor. This way, the disposal of actions is
performed in a more consistent manner (similar for view and editor).