[NONE] Use bulk changes to reduce number of fired events
The GraphCopier uses bulk operations on collections when copied nodes
and edges are added to the resulting graph. This greatly helps to reduce
the number of fired events and therefore improves the runtime performance.

The same pattern has been applied in the ZestContentViewer when the graph
is created from the IGraphContentProvider.

Signed-off-by: Sebastian Zarnekow <sebastian.zarnekow@gmail.com>